Hw5: rename link_update_searching and exception_handler packages
