+++ scrapper/src/main/java/edu/java/configuration/LinkUpdateSearchingConfiguration.java
@@ -3,9 +3,9 @@ package edu.java.configuration;
 import edu.java.client.github.GithubClient;
 import edu.java.client.stackoverflow.StackoverflowClient;
 import edu.java.service.LinkService;
-import edu.java.service.linkupdatesearching.searchers.LinkUpdateSearcher;
-import edu.java.service.linkupdatesearching.searchers.github.GithubLinkUpdateSearcher;
-import edu.java.service.linkupdatesearching.searchers.stackoverflow.StackOverFlowLinkUpdateSearcher;
+import edu.java.service.link_update_searching.searchers.LinkUpdateSearcher;
+import edu.java.service.link_update_searching.searchers.github.GithubLinkUpdateSearcher;
+import edu.java.service.link_update_searching.searchers.stackoverflow.StackOverFlowLinkUpdateSearcher;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
